coala_main: Move show_bears logic to coala.py
The show_bears options (-A and -B) are only valid for the
coala cli based binary. Hence, we do not need it in
coala_main which is a common entry point for all binaries.
This commit moves the logic into coala. This also allows us
to not parse all the sections for bears unnecessarily for
show_all_bears option.

Related to https://github.com/coala-analyzer/coala/issues/1634