coalaTest: Fix show_all_bears test
The earlier test for show-all-bears was not very good, as it
was just counting the number of lines >= 0. Now that we have
bear_test_module() we can actually check if the number of bears
found is the same as the number of bears in the folder.

Fixes https://github.com/coala-analyzer/coala/issues/1664